修改bug
修改 多次调用sql会造成sql错乱的bug

+++ src/main/java/top/hejiaxuan/util/maker/delete/DefaultDelete.java
@@ -12,15 +12,22 @@ public class DefaultDelete extends SqlWhere implements Delete {
     @Override
     public String toSql() {
         Assert.isTrue(sqlWhere.length() != 0, "没有添加条件");
+        if (completeSql != null) {
+            return completeSql;
+        }
         sql.append(StringUtils.append("DELETE FROM ", tableName, " WHERE ", sqlWhere));
-        String sqlStr = sql.toString();
-        return sqlStr;
+        completeSql = sql.toString();
+        return completeSql;
     }
 
     @Override
     public Object[] getSqlValues() {
+        if (completeSqlValues != null) {
+            return completeSqlValues;
+        }
         sqlValues.addAll(super.whereValues);
-        return sqlValues.toArray();
+        completeSqlValues = sqlValues.toArray();
+        return completeSqlValues;
     }
 
 }
